+++ {{cookiecutter.github_repo_name}}/terraform/modules/eks/test-app2.tf
@@ -1,21 +1,7 @@
 
-# nginx deployment example: https://registry.terraform.io/providers/hashicorp/kubernetes/latest/docs/guides/getting-started
-
-resource "kubernetes_namespace" "app" {
-  metadata {
-    labels = {
-      app = "owncloud"
-    }
-    name = "fargate-node"
-  }
-
-  depends_on = [module.eks]
-}
-
 resource "kubernetes_deployment" "app" {
   metadata {
-    name      = "owncloud-server"
-    namespace = kubernetes_namespace.app.metadata.0.name
+    name = "owncloud-server"
     labels = {
       app = "owncloud"
     }
@@ -49,14 +35,12 @@ resource "kubernetes_deployment" "app" {
       }
     }
   }
-  depends_on = [kubernetes_namespace.app]
 
 }
 
 resource "kubernetes_service" "app" {
   metadata {
-    name      = "owncloud-service"
-    namespace = kubernetes_namespace.app.metadata.0.name
+    name = "owncloud-service"
   }
   spec {
     selector = {
@@ -77,8 +61,7 @@ resource "kubernetes_service" "app" {
 
 resource "kubernetes_ingress" "app" {
   metadata {
-    name      = "owncloud-lb"
-    namespace = kubernetes_namespace.app.metadata.0.name
+    name = "owncloud-lb"
     #annotations = {
     #  "kubernetes.io/ingress.class"           = "alb"
     #  "alb.ingress.kubernetes.io/scheme"      = "internet-facing"